Update GitHub Pages build and deployment process
Modified `jekyll-gh-pages.yml` to enhance the build and deployment process for documentation and GitHub Pages. Updated the "Copy doxygen files" step to copy files to `./public/html`. Added a new `build-page` job to build the site with Jekyll and upload artifacts. Introduced a `deploy` job that depends on `build-page` to deploy to GitHub Pages.